twitch_tools/requirements.txt
    # Twitch API Tools Requirements

# Основные библиотеки для работы с Twitch API
twitchAPI>=4.5.0
aiohttp>=3.9.3
asyncio>=3.4.3

# Работа с конфигурацией
python-dotenv>=1.1.0

# HTTP запросы
requests>=2.32.0

# Дополнительные утилиты
typing_extensions>=4.0.0
python-dateutil>=2.8.0
